requirements.txt
aws-cdk-lib==2.64.0
constructs>=10.0.0,<11.0.0
cdk-nag>=2.21.0
